GenreWordConverterをインターフェースに区切って依存関係を明確に
